ddl_bronze
This script creates tables in the 'bronze' schema, dropping existing tables. Run this script to re-define the DDL structure of 'bronze' tables.

Important Note:
This script is designed for use with PostgreSQL. Make sure you are connected to the correct database before running it, as existing data in these tables will be permanently deleted.